backend/pyproject.toml
[tool.poetry]
name = "insurance_agency"
version = "0.1.0"
description = ""
authors = ["azat715 <azat715@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pylint = "^2.9.3"
pylint-django = "^2.4.4"
pytest = "^6.2.4"
isort = "^5.9.2"
ipython = "^7.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
